Added messagefactory to node
removed unused imports
