+++ src/main/java/nl/erwinvaneyk/communication/Message.java
@@ -3,7 +3,6 @@ package nl.erwinvaneyk.communication;
 import java.io.Serializable;
 import java.util.Date;
 
-import lombok.NonNull;
 import nl.erwinvaneyk.core.NodeAddress;
 
 public interface Message extends Serializable {
