+++ src/main/java/nl/erwinvaneyk/core/NodeImpl.java
@@ -2,31 +2,40 @@ package nl.erwinvaneyk.core;
 
 import java.rmi.RemoteException;
 
-import lombok.Data;
+import lombok.Getter;
+import lombok.ToString;
+import nl.erwinvaneyk.communication.BasicMessageFactory;
 import nl.erwinvaneyk.communication.Connector;
 import nl.erwinvaneyk.communication.ConnectorImpl;
-import nl.erwinvaneyk.communication.exceptions.CommunicationException;
-import nl.erwinvaneyk.communication.rmi.NoBufferMessageHandler;
-import nl.erwinvaneyk.communication.Server;
 import nl.erwinvaneyk.communication.MessageDistributor;
+import nl.erwinvaneyk.communication.MessageFactory;
+import nl.erwinvaneyk.communication.Server;
+import nl.erwinvaneyk.communication.exceptions.CommunicationException;
 import nl.erwinvaneyk.communication.handlers.NodeConnectHandler;
 import nl.erwinvaneyk.communication.handlers.NodeInitHandler;
+import nl.erwinvaneyk.communication.rmi.NoBufferMessageHandler;
 
-@Data
+@ToString
 public class NodeImpl implements Node {
 
 	public static final String NODE_TYPE = NodeImpl.class.getSimpleName();
 
+	@Getter
 	private final NodeState state;
-
+	@Getter
 	private final MessageDistributor messageDistributor;
-	private final Server registry;
+	@Getter
 	private final Connector connector;
+	@Getter
+	private final MessageFactory messageFactory;
+
+	private final Server server;
 
 	protected NodeImpl(NodeAddress address, Server registry, String clusterId) throws CommunicationException {
 		this.state = new NodeState(address, clusterId);
-		this.registry = registry;
+		this.server = registry;
 		this.connector = new ConnectorImpl(this.getState());
+		this.messageFactory = new BasicMessageFactory(address);
 		try {
 			messageDistributor = new NoBufferMessageHandler(address);
 			messageDistributor.bind(new NodeInitHandler(connector, state));
@@ -35,7 +44,7 @@ public class NodeImpl implements Node {
 		catch (RemoteException e) {
 			throw new CommunicationException("Failed to launch MessageHandler", e);
 		}
-		// Setup registry
+		// Setup server
 		registry.start(address.getLocation().getPort());
 		registry.register(address, messageDistributor);
 	}
@@ -47,6 +56,6 @@ public class NodeImpl implements Node {
 
 	@Override
 	public void disconnect() {
-		registry.shutdown();
+		server.shutdown();
 	}
 }
