+++ src/test/java/nl/erwinvaneyk/communication/rmi/RMISocketTest.java
@@ -1,6 +1,7 @@
 package nl.erwinvaneyk.communication.rmi;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 import java.net.UnknownHostException;
 import java.rmi.RemoteException;
